refactored the exercise instance page
