+++ src/util/weightUnitConverting.ts
@@ -17,13 +17,18 @@ export const convertKgsToLbs = (kgs: number): number => {
 };
 
 export const convertLbsToKgs = (lbs: number): number => {
-  const kgs = lbs / 2.20462;
-  const integerPart = Math.floor(kgs);
-  const decimalPart = kgs - integerPart;
+  return lbs / 2.20462;
+  // const kgs = lbs / 2.20462;
+  // const integerPart = Math.floor(kgs);
+  // const decimalPart = kgs - integerPart;
 
-  return decimalPart < 0.5
-    ? integerPart
-    : decimalPart === 0.5
-    ? integerPart + 0.5
-    : integerPart + 1;
+  // return decimalPart < 0.5
+  //   ? integerPart
+  //   : decimalPart === 0.5
+  //   ? integerPart + 0.5
+  //   : integerPart + 1;
+};
+
+export const roundKgs = (kgs: number): number => {
+  return parseFloat(kgs.toFixed(2));
 };
